+++ app/salaries/routes/__init__.py
@@ -1,3 +1,4 @@
 from .salaries import router as salaries_router
 from .technologies import router as technologies_router
 from .titles import router as titles_router
+from .seniority import router as seniority_router
