+++ app/salaries/routes/seniority.py
@@ -1,33 +1,31 @@
 # Python
 from typing import List
-
 # FastAPI
 from fastapi import APIRouter, Body
 from fastapi import status
 
 # Project
-from app.salaries.mockdata.salary_mockdata import all_technologies
-
+from app.salaries.mockdata.salary_mockdata import all_seniority
 
 router: APIRouter = APIRouter()
 
 
 @router.get(
-    path="/api/technologies",
+    path="/api/seniority",
     status_code=status.HTTP_200_OK,
-    summary="Get technologies availables.",
+    summary="Get levels availables.",
     tags=["Salaries"],
     response_model=List,
 )
-def technologies():
+def titles():
     """
-    # Technologies
-    Get technologies available in our data set.
+    # Seniority
+    Get levels availables.
 
     # Parameters
     - None
 
     # Return
-        List of all available technologies.
+        List of all available levels.
     """
-    return all_technologies()
+    return all_seniority()
