--- app/salaries/schemas/__init__.py
@@ -1,1 +1,0 @@
-from .salaries import Salary, SalaryOut
