--- app/salaries/schemas/technologies.py
@@ -1,10 +1,0 @@
-
-# Python
-from typing import List
-
-# Pydantic
-from pydantic import BaseModel, Field
-
-
-class Technology(BaseModel):
-    name: str = Field(..., example="python")
